This commit adds a hashing/crypto dependency for another security measure in development. Below is a list of changes:
package.json:
	- Added node_hash v0.2.0 as a dependency

util/common/security.js:
	- Fixed a typo involving a missing semicolon

util/route_handlers.js:
	- Added a temporary piece of code that implements the node_hash.sha256 algorithm to test out the hashing. I plan to have all session IDs hashed by using the current ISO string timestamp from JavaScript Date.toISOString(), combined with the submitted username. This sessionID will then be sent to the sessionData collection in the database before being sent over to the client as a redirection order (maybe a code 301 with the full url includeing the sessionID as a querystring parameter?)

Note: With the addition of the node_hash function, it is imperative that I create a module (shall I call it "hasher.js"?) that standardizes the way we create sessionIDs, password hashes, and salts. Then, I must update the database schema and setup script such that each user has a unique salt. Finally, I must update the common resources readme to tell the user that the syskey.passWord field of the credentials.json object MUST be the already-encrypted version of the password, and that generation of this enrypted password should be acquired to the appropriate function in "hasher.js".