This commit streamlines the MDBI system by merging the projection fun… (#19)
* This commit streamlines the MDBI system by merging the projection functionality to the "findDocs()" MongoDB API Wrapper function, eliminating the need for "findAndProjectDocs()".
Below is a list of changes:

mdbi/mdbiRoutes/index.js:
	- Modified the logic of the "mdbi/search/documents" endpoint handler to implement projections with the "findDocs()" function
	- Removed all implementation of the "findAndProjectDocs()" function

mdbi/mongoWrapper.js:
	- Modified the documentation of the "findDocs()" API wrapper to describe the "projection" parameter implementation
	- Removed the "findAndProjectDocs()" API wrapper, as its presence is redundant now that the "findDocs()" function allows projections

* This commit adds "mdb.findDocCount()", which is better suited for finding the count of result sets than the "mdb.findDocs()" function.
However, it still is not implemented; I forsee a need to change the way data is passed to the client side, if I end up using this function. This would be done either by modifying the response body of the "mdbi/search/documents" endpoint, or by creating an entirely new endpoint for just the count.