This commit adds the "/api/user/edit" endpoint, which combines the functions of various legacy endpoints to modify a user. In an effort to reduce the length of the endpoint, utility functions "organizeUpdates()" and "buildEditQueries()" were added to abstract away some of the misc logic, and kept the endpoint's focus on the logic of the user editing operations.
This commit also adds various new utility features, such as the ability to put API handlers on
a blacklist, thereby causing the logger module to ignore those messages (this still needs to
be tested). The function "api_util.verifySession" was also modified to give information on the
session that was returned from the search, new mdbi error format objects were defined, and a
new response format object was added, which will serve as a general template for responses to
the client.