This commit fixes the logic for the "/api/user/edit" API, which can now facilitate the consecutive modification of two collections "Member" and "MembershipData" via Promises. Response status codes were also changed to better reflect the status of the server's reception of the client's message (as their sole responsibility should be).
In an effort to give the client a more detailed description of the effects of their
edit queries, a new error_format option "MDBI_PARTIAL_EFFECT" was added and used in
cases where only some of the intended collections were modified.

On a side note, response_formats.js was also updated with a new handler tag for
".asCommonStr()" and the logic for comparing template names against existing
templates was corrected.