Rj/membership application (#60)
* Initial commit of the rj/membershipApplication branch,

which adds the router logic, navbar dropdown link and initial jsx
of the membership application component.

* Added preliminary fields and field capturing logic

for React.

* Adds request logic to pull the availability of a specific username.

However, this requires supplementary APIs to promote secure search
of usernames and info

* Adds initial public APIs for Membership Application,

namely to check if a username is available. Also added changes to
the .gitignore to ignore certain VS Code residual files.

* Integrated Membership Application Username Check API to front-end

* Adds membership application form front-end component

and accompanying back-end logic to support this.