Updated TravisCI Script (#117)
* Updated TravisCI Script

Now Travis will run tests and build the application before a pull
request can be merged. I also moved the existing script to a depricated
folder for future reference.

* Added Function to Close MongoDB Connection

This is done to eliminate the possiblity of any hanging connections
as a result of running test cases. Mocha runs tests in alphabetical
order, so the function is called only in users.js, since it will be run
last.

* changed test directory

* added alt text

* Added conditional to Server.js to retain behavior

Before, running node server.js will have the server listen for requests
indefinitely. The refactor took away this functionality as all the code
was placed into a class. Now, running node server.js will keep its old
behavior from the if conditional.