Refactored Routing + Navbar Rendering (#171)
* Refactored Routing

We now check for the user's auth status in the highest component and use a
custom PrivateRoute component for securing routes. We also now use array.map in a seperate file, Routing.js to render the routes

* Refactored NavBar Rendering

Now, each component in Routing.js is wrapped by the NavbarWrapper
component. NavbarWrapper renders the correct component based on the
route and user's account priviledge level.