Run github actions on merge
Similar to [SCE-discord-bot](https://github.com/SCE-Development/SCE-discord-bot), it would be nice to have a green check to show off our code on dev as passing. As of now, we only run the GitHub Actions build when a PR is opened. This changes it to also have the build be ran when dev is updated.