Merge multiple PRs into a single PR (#954)
* Run github actions on merge

Similar to [SCE-discord-bot](https://github.com/SCE-Development/SCE-discord-bot), it would be nice to have a green check to show off our code on dev as passing. As of now, we only run the GitHub Actions build when a PR is opened. This changes it to also have the build be ran when dev is updated.

* removed space around dev

* put spaces back

* moved token-functions to util folder

* i messed up

* updated paths +mocha tests pass

* grrrrr

* GRRRRR

* just let me please

* Passes all API tests. Reset package-lock + package to dev branch.

* npm audit fix

* removed space around dev

* put spaces back

* moved token-functions to util folder

* i messed up

* npm audit fix

* Added combine prs workflow

* removed space around dev

* put spaces back

* moved token-functions to util folder

* i messed up

* fixed package-lock.json

* newline at end of package-lock.json

Co-authored-by: Allen Tran <allentran@allens-macbook.local>