fixed css bug caused by homepage image frame
There used to be a bug on the homepage that messed up the whole frontend for mobile devices
Ideally when sliding horizontally, CSS will take care of the logic and will not show the whitespace underneath
But the homepage image frame was setting a margin which pushed it off of the screen causing it to show the white space when sliding horizontally. I fixed it by adding padding instead of a margin.

p.s. most was just trial and error 💀